[confcom] Updating katapolicygen to support new node image (#7849)
* changing containers to be hosted on mcr

* changing input args for genpolicy

* some containers use empty string env vars. this allows for that

* updating tests and readme for new kata interface

* updating images and getting rid of unused code

* moving import statement

* using MCR images

* updating test to have empty command

+++ src/confcom/azext_confcom/tests/latest/test_confcom_kata.py
@@ -5,7 +5,6 @@
 
 import os
 import unittest
-import pytest
 from azext_confcom.custom import katapolicygen_confcom
 
 import pytest
@@ -49,22 +48,39 @@ spec:
 
     def test_invalid_config_map_path(self):
         filename = "pod.yaml"
-        with open(filename, "w") as f:
-            f.write(KataPolicyGen.pod_string)
-        with self.assertRaises(SystemExit) as wrapped_exit:
-            katapolicygen_confcom(
-                filename, "fakepath/configmap.yaml",
-            )
-        os.remove(filename)
+        try:
+            with open(filename, "w") as f:
+                f.write(KataPolicyGen.pod_string)
+            with self.assertRaises(SystemExit) as wrapped_exit:
+                katapolicygen_confcom(
+                    filename, "fakepath/configmap.yaml",
+                )
+        finally:
+            if os.path.exists(filename):
+                os.remove(filename)
         self.assertNotEqual(wrapped_exit.exception.code, 0)
 
-    def test_invalid_settings(self):
+    def test_valid_settings(self):
         filename = "pod2.yaml"
-        with open(filename, "w") as f:
-            f.write(KataPolicyGen.pod_string)
+        try:
+            with open(filename, "w") as f:
+                f.write(KataPolicyGen.pod_string)
+            with self.assertRaises(SystemExit) as wrapped_exit:
+                katapolicygen_confcom(
+                    filename, None
+                )
+            with open(filename, "r") as f:
+                content = f.read()
+        finally:
+            if os.path.exists(filename):
+                os.remove(filename)
+        self.assertEqual(wrapped_exit.exception.code, 0, "Policy not generated successfully")
+        self.assertNotEqual(content, KataPolicyGen.pod_string, "Policy content not changed in yaml")
+
+    def test_print_version(self):
         with self.assertRaises(SystemExit) as wrapped_exit:
             katapolicygen_confcom(
-                filename, None, settings_file_name="genpolicy-settings.json"
+                None, None, print_version=True
             )
-        os.remove(filename)
-        self.assertEqual(wrapped_exit.exception.code, 1)
+
+        self.assertEqual(wrapped_exit.exception.code, 0)
